% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compare_models}
\alias{compare_models}
\title{compare_models}
\usage{
compare_models(city = "nyc", from = TRUE, measure = "covar",
  plot = FALSE, data.dir = "./results/")
}
\arguments{
\item{city}{nyc, washington, chicago, boston, london (case insensitive)}

\item{from}{TRUE or FALSE}

\item{measure}{can be ('covar', 'ntrips', 'info'), where the latter is mutual
information.}

\item{plot}{= TRUE allows visual inspection of individual station results}
}
\value{
a list of data frames, one for each model, containg data frame with
rows for each station and columns of [SS, k, b, AIC], where k is the width
parameter of the model, and b is the shape parameter (which does not apply to
power-law, Gaussian or Cauchy models).
}
\description{
Statistical comparisons of 6 different functional forms of distance decay:
power-law, Gaussian, exponential, Weibull, Cauchy, and Box-Cox
}

