% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fit_gaussian}
\alias{fit_gaussian}
\title{fit_gaussian}
\usage{
fit_gaussian(city = "nyc", from = TRUE, measure = "covar", std = TRUE,
  nearfar = 0, subscriber = 0, mf = 0, msg = FALSE,
  data.dir = "../results/")
}
\arguments{
\item{city}{nyc, washington, chicago, boston, london (case insensitive)}

\item{from}{TRUE or FALSE}

\item{measure}{can be ('covar', 'ntrips', 'info'), where the latter is mutual
information.}

\item{std}{is not implemented}

\item{nearfar}{= (0,1,2) for (all, near, far)}

\item{subscriber}{= 0 to analyse all data; 1 to analyse subscribers only; 2
to analyse non-subscribers only; 3 to perform aged-based analyses}

\item{mf}{= 1/2 for male/female data (for subscriber=1 only)}

\item{data.dir}{= directory containing results from the C++ routines}
}
\value{
a data frame of 8 columns (i,k,y,ntrips,ss,r2,resid.slope,f), where i
is station number, k is variance of Gaussian fit, resid.slope is measured
against distance, and f is the corresponding f-statistic
}
\description{
Produces a data frame with [i, k, y, ss, ntrips, F], where F is an
F-statistic for changes in the variance of model residuals with distance.
This is calculated by binning them into 20 bins, and so can be tested with
p = 1 - pf (mean (f), 1, 18). Note, however, that slopes of these
relationships can differ in sign, so perhaps a more robust way to test is
simply with a T-test on the distribution of slopes of these residual
regressions.
}
\details{
Curves may not be able to be fitted to every station, and results include
only those stations for which fits were possible. Statistics for residuals
can be evaluated with "test.resids".
}

