% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_bike_data}
\alias{get_bike_data}
\title{get_bike_data}
\usage{
get_bike_data(city = "nyc", from = TRUE, measure = "covar", std = TRUE,
  nearfar = 0, subscriber = 0, mf = 0, data.dir = "../../results/",
  msg = FALSE)
}
\arguments{
\item{city}{nyc, washington, chicago, boston, london (case insensitive)}

\item{from}{TRUE or FALSE}

\item{measure}{can be ('covar', 'ntrips', 'info'), where the latter is mutual
information.}

\item{std}{is not implemented}

\item{nearfar}{= (0,1,2) for (all, near, far)}

\item{subscriber}{= 0 to analyse all data; 1 to analyse subscribers only; 2
to analyse non-subscribers only; 3 to perform aged-based analyses}

\item{mf}{= 1/2 for male/female data (for subscriber=1 only)}

\item{data.dir}{= directory containing results from the C++ routines}
}
\value{
a list of 2 matrices of d=inter-station distances, y=value according
to measure parameter, along with xy=lat-lon coordinates of stations.
}
\description{
Loads hire bicycle data pre-generated by the C++ routines in
github.com/mpadge/bike-correlations.
}

