% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_vgram}
\alias{plot_vgram}
\title{plot_vgram}
\usage{
plot_vgram(city = "nyc", measure = "covar", nearfar = 0, subscriber = 0,
  mf = 0, data.dir = "results/")
}
\arguments{
\item{city}{nyc, washington, chicago, boston, london (case insensitive)}

\item{from}{TRUE or FALSE}

\item{measure}{can be ('covar', 'ntrips', 'info'), where the latter is mutual
information.}

\item{std}{is not implemented}

\item{nearfar}{= (0,1,2) for (all, near, far)}

\item{subscriber}{= 0 to analyse all data; 1 to analyse subscribers only; 2
to analyse non-subscribers only; 3 to perform aged-based analyses}

\item{mf}{= 1/2 for male/female data (for subscriber=1 only)}

\item{data.dir}{= directory containing results from the C++ routines}
}
\description{
Plots spatial semivariogram of k-values.
}
\details{
Distances are in units of degrees, not kilometres.
}

